refactor(node): extract balance / history / genesis-wallets to commands/misc.rs
Three top-level helpers move out of main.rs:
  cmd_balance, cmd_history — read-only account queries
  cmd_genesis_wallets — one-shot bootstrap helper that generates the
                        seven labelled keypairs (founder, ecosystem_fund,
                        early_validator, reserve, genesis_node_1..3) and
                        writes them to <wallets_dir>/genesis_wallets.json

Different domains but small enough that one misc module is cleaner than
splintering. balance/history hit Blockchain accessors; genesis_wallets
generates wallets + writes a plaintext file (the CRITICAL backup-and-delete
reminder is printed by the command).

main.rs drops to 3890 lines.
